// src/world/transfers/generated-routes.ts
/* Generated by scripts/content/build-map-v2.ts. */
export const GENERATED_NEIGHBORHOOD_ROUTES = [
  {
    "originMapId": "northeast_downtown",
    "destinationMapId": "northwest_residential",
    "sourcePortalId": "from-residential",
    "sourcePortalTile": {
      "x": 0,
      "y": 24
    },
    "destinationEntranceId": "to-downtown",
    "destinationEntranceTile": {
      "x": 63,
      "y": 24
    }
  },
  {
    "originMapId": "northeast_downtown",
    "destinationMapId": "southeast_docks",
    "sourcePortalId": "to-docks",
    "sourcePortalTile": {
      "x": 32,
      "y": 47
    },
    "destinationEntranceId": "from-downtown",
    "destinationEntranceTile": {
      "x": 32,
      "y": 0
    }
  },
  {
    "originMapId": "northwest_residential",
    "destinationMapId": "southwest_commercial",
    "sourcePortalId": "to-commercial",
    "sourcePortalTile": {
      "x": 32,
      "y": 47
    },
    "destinationEntranceId": "from-residential",
    "destinationEntranceTile": {
      "x": 32,
      "y": 0
    }
  },
  {
    "originMapId": "northwest_residential",
    "destinationMapId": "northeast_downtown",
    "sourcePortalId": "to-downtown",
    "sourcePortalTile": {
      "x": 63,
      "y": 24
    },
    "destinationEntranceId": "from-residential",
    "destinationEntranceTile": {
      "x": 0,
      "y": 24
    }
  },
  {
    "originMapId": "southeast_docks",
    "destinationMapId": "southwest_commercial",
    "sourcePortalId": "from-commercial",
    "sourcePortalTile": {
      "x": 0,
      "y": 24
    },
    "destinationEntranceId": "to-docks",
    "destinationEntranceTile": {
      "x": 63,
      "y": 24
    }
  },
  {
    "originMapId": "southeast_docks",
    "destinationMapId": "northeast_downtown",
    "sourcePortalId": "from-downtown",
    "sourcePortalTile": {
      "x": 32,
      "y": 0
    },
    "destinationEntranceId": "to-docks",
    "destinationEntranceTile": {
      "x": 32,
      "y": 47
    }
  },
  {
    "originMapId": "southwest_commercial",
    "destinationMapId": "northwest_residential",
    "sourcePortalId": "from-residential",
    "sourcePortalTile": {
      "x": 32,
      "y": 0
    },
    "destinationEntranceId": "to-commercial",
    "destinationEntranceTile": {
      "x": 32,
      "y": 47
    }
  },
  {
    "originMapId": "southwest_commercial",
    "destinationMapId": "southeast_docks",
    "sourcePortalId": "to-docks",
    "sourcePortalTile": {
      "x": 63,
      "y": 24
    },
    "destinationEntranceId": "from-commercial",
    "destinationEntranceTile": {
      "x": 0,
      "y": 24
    }
  }
] as const;